Fix origin for warnings due to event methods (#100500)
This fixes the warning origin to prevent warnings from being
reported on events for annotated event methods. Instead the
warning is shown at the location that introduces the dependency
on the event.

Doesn't fix the multiple warnings reported in some cases. This is
due to MarkEvent also marking event methods, unlike
MarkProperty. We might consider fixing this too, but it was
looking like a larger change than I wanted to make here.

This removes some spurious warnings that were showing up when
bubbling up RUC in winforms.