Make InstallRuntimeComponentToFinalDestination use NativeOutputPath for copying to runtime output (#100782)
The InstallRuntimeComponentToFinalDestination target from infrastructure for NativeAOT-ed runtime components (https://github.com/dotnet/runtime/pull/98565) copies the component and symbol files to the runtime output directory. Make the target use NativeOutputPath for the for the library and symbol path instead of the ones copied to the publish directory. They should be available after LinkNative.

This was working in the runtime repo because the projects under src/native/managed are currently using the ILCompiler package corresponding to the SDK being used to build the repo. That version of the package doesn't have https://github.com/dotnet/runtime/pull/98342 - the native-library.targets were relying the symbols already being copied before _CopyAotSymbols was run.